requirements.txt
python-dotenv
azure-cognitiveservices-vision-customvision